Fork the gem under a different name
This may be temporary, but we are going to publish our fork
to rubygems under another name.

This will allow us to include this version as a dependency in the
govuk_tech_docs gem we publish here:
https://rubygems.org/gems/govuk_tech_docs

The difference between this and upstream is a change from therubyracer
to execjs. This fixes problems with middleman hanging in dev mode.
See https://github.com/manastech/middleman-search/pull/24 for more
details.