Allow 'content' field to be overridden
We noticed that the search index performance degraded noticably
when [adding lots of documents to the GOV.UK Dev Docs][search-pr].
This meant that the 'Loading search index' message appeared for
longer than previously.

The reason for this slowness was that the new documents were
large, and including all of their contents made the generated
search.json file bigger and slower to load into memory.

In reality, we only really want to index the titles of these
pages, to make the pages findable. We don't particularly want
to index their contents, as they're not useful search terms.

For example, the [Document type: about] page is merely a
'landing page' summarising how often the document type is used,
and by which apps. Others, such as [Schema: detailed_guide]
are packed with very 'internal' information that is unlikely to
be searched for in isolation.

This commit allows upstream projects to specify their own `content`
value, rather than it always being the generated HTML of the
document. This means we can pass `content: ""` where indexing
page contents is unnecessary, allowing us to add them to the
search index without taking the performance hit.

[search-pr]: https://github.com/alphagov/govuk-developer-docs/pull/2509
[Document type: about]: https://docs.publishing.service.gov.uk/document-types/about.html
[Schema: detailed_guide]: https://docs.publishing.service.gov.uk/content-schemas/detailed_guide.html