Version bump to 0.11.1
This is a new feature, so best treated as a minor (rather than
patch) release. Given that this gem isn't using semantic
versioning, the middle number probably takes the position of
'major', so the final number takes the position of 'minor'.

I don't know why the previous version was set to 0.11.0a and not
0.11.0 (the origin version was 0.10.0) - that suggests both a
major change and a patch, if we take the 'a' to mean 'patch'. The
alphanumeric syntax is one I've only come across in `.ruby-version`
files, so I hope my bump to 0.11.1 isn't too controversial!

This is an internal library and unofficial fork, so I don't think
it matters too much what version we end up naming.