Rename middleman-search.gemspec -> middleman-search-gds.gemspec
In https://github.com/alphagov/middleman-search/pull/8, we pass
the gem name to govuk-jenkinslib so that it no longer uses the
repo name (by default), which would be 'middleman-search' rather
than 'middleman-search-gds'.

However, this alone doesn't fix the problem, as Jenkins then looks
for a 'middleman-search-gds.gemspec' file. Ours is called
'middleman-search.gemspec', probably just because we forked from
middleman-search and never renamed the file (despite renaming the
'name' property inside that file). This commit corrects that.